% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_parcels.R
\name{categorize_parcels}
\alias{categorize_parcels}
\title{categorize_parcels}
\usage{
categorize_parcels(data, dictionary)
}
\arguments{
\item{data}{a data frame with the columns "parcel_number" (e.g., sentence, paragraph, message, words spoken during a timeframe, etc) and "word" (indivdiual words stripped of non alphanumeric characters). I.e., the output from `tidy_parcels()`.}

\item{dictionary}{a data frame with the columns "category" and "word", representing the specific words that are assigned to superordinate categories. Notionally, a given word could be a member of multiple categories.}
}
\value{
a data frame with the columns "parcel_number", "category", "count" (instances of words per category in each parcel), and "percent" (instances of words from this category as percentage of all words in the parcel).
}
\description{
This function takes tokenized text (e.g., the output from `tidy_parcels()`) and a applies your desired dictionary/lexicon of word categories to it. The frequency of words per category is calculatd along with the percentage of word per category per total words in the text parcel.
Three dictionaries are included: emotions (8 emotions), valence (positive vs. negative words) and relations (relational terms). The latter is in development. Any custom dictionary that includes the columns "category" and "word" can be used.
}
\examples{
categorized_data <- reddit_suicide_data \%>\%
    tidy_parcels() \%>\%
    categorize_parcels(dictionary = relations)
}
