% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotions.R
\docType{data}
\name{emotions}
\alias{emotions}
\title{emotions dictionary}
\format{A data frame frame with the columns "category" and "word".}
\usage{
emotions
}
\description{
A dictionary/lexicon of emotional words in 8 non-exclusive categories: anticipation, anger, disgust, fear, joy, sadness, surprise, trust. Based on the "nrc" lexicon from the `tidytext` package, but excludes the positive and negative categories.
}
\keyword{datasets}
