% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_frequencies.R
\name{plot_frequencies}
\alias{plot_frequencies}
\title{plot_frequencies}
\usage{
plot_frequencies(data)
}
\arguments{
\item{data}{a data frame with the columns "parcel_number", "category" and "count".}
}
\value{
a ggplot plot object.
}
\description{
Plots raw data, distributions, and means with 95% CIs. Takes output from `categorize_parcels()` and plots frequencies by dictionary categories.
}
\examples{
categorized_data <- reddit_suicide_data \%>\%
    tidy_parcels() \%>\%
    categorize_parcels(dictionary = relations)
plot_frequencies(data = categorized_data)
}
