% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_percentages.R
\name{plot_percentages}
\alias{plot_percentages}
\title{plot_percentages}
\usage{
plot_percentages(data)
}
\arguments{
\item{data}{a data frame with the columns "parcel_number", "category" and "percent".}
}
\value{
a ggplot plot object.
}
\description{
Plots raw data, distributions, and means with 95% CIs. Takes output from `categorize_parcels()` and plots percentages of total words by dictionary categories.
}
\examples{
categorized_data <- reddit_suicide_data \%>\%
    tidy_parcels() \%>\%
    categorize_parcels(dictionary = relations)
plot_percentages(data = categorized_data)
}
