% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_suicide_data.R
\docType{data}
\name{reddit_suicide_data}
\alias{reddit_suicide_data}
\title{reddit_suicide_data}
\format{A data frame frame with the columns "category" and "word".}
\source{
Data taken from an askreddit thread titled "Suicide survivors of Reddit, what was your first conscious thought after you realized that you hadn't succeeded? Serious Replies Only" (https://www.reddit.com/r/AskReddit/comments/4e8oip/serious_suicide_survivors_of_reddit_what_was_your/?sort=top). This includes c.2000 comments. Data was scraped using Reddit Comment Scraper (https://github.com/ianhussey/RedditCommentScraper) with some semi-automated processing in Excel.
}
\usage{
reddit_suicide_data
}
\description{
A dataset of natural language comments scraped from an reddit.com/r/askreddit thread on immediate reactions to surviving a suicide attempt.
}
\keyword{datasets}
