% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\docType{data}
\name{relations}
\alias{relations}
\title{Relations dictionary}
\format{A data frame frame with the columns "category" and "word".}
\usage{
relations
}
\description{
A dictionary or lexicon containing synonyms of relational terms.
}
\keyword{datasets}
