% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP.R
\docType{package}
\name{simpleNLP-package}
\alias{simpleNLP}
\alias{simpleNLP-package}
\title{Tokenize, process, and visualise text data using Natural Language Processing}
\description{
Simple workflows build on the tidytext package. Multiple dictionaries/lexicons are provided: \code{\link{emotions}}, \code{\link{valence}}, and \code{\link{relations}}. Plotting is done via ggplot2.
}
\details{
Use \code{\link{tidy_parcels}} to tokenize data, then \code{\link{categorize_parcels}} to process it using a dictionary. Finally, plot it with \code{\link{plot_percentages}} before analysis.
}
\author{
\strong{Maintainer}: Ian Hussey \email{ian.hussey@ugent.be}

}
