% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_parcels.R
\name{tidy_parcels}
\alias{tidy_parcels}
\title{tidy_parcels}
\usage{
tidy_parcels(data)
}
\arguments{
\item{data}{a data frame with the columns "id" (e.g., individiual, text source, etc) and "parcel" (e.g., sentence, paragraph, message, words spoken during a timeframe, etc).}
}
\value{
a data frame with the columns "id", "parcel_number" (incremental integer assigned to the parcels) and "word" (individual words extracted from each parcel).
}
\description{
This function takes text, removes all alphanumeric characters, and tokenizes it for natural language processing. I.e., it returns a data frame with one word per row, and the parcel number that the word came from. Parcels could be sentences, pragraphcs, timespans, etc.
}
\examples{
tidied_data <- tidy_parcels(data = reddit_suicide_data)
}
