% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valence.R
\docType{data}
\name{valence}
\alias{valence}
\title{valence dictionary}
\format{A data frame frame with the columns "category" and "word".}
\usage{
valence
}
\description{
A dictionary/lexicon of positive vs negatively valenced words. Based on the "bing" lexicon from the `tidytext` package.
}
\keyword{datasets}
